Fixed scalajs stackoverflow issue
